// jakuv-card-game_updated/types.ts
export enum Suit {
  Spades = '♠',
  Clubs = '♣',
  Hearts = '♥',
  Diamonds = '♦',
}

export enum CardColor {
  Red = 'Red',
  Black = 'Black',
}

export type Rank = 'A' | '2' | '3' | '4' | '5' | '6' | '7' | '8' | '9' | '10' | 'J' | 'Q' | 'K' | '?';

export type AceValue = 1 | 3 | 5;

export interface Card {
  id: string;
  rank: Rank;
  suit: Suit;
  color: CardColor;
  isFaceUp: boolean;
  aceValue?: AceValue;
  oneTimeProtection?: boolean;
}